src/drift/orchestrator.py
from absl import app
from absl import flags
from absl import logging
import jax.numpy as jnp

FLAGS = flags.FLAGS
flags.DEFINE_float("t1_decay_mean", 30.0, "Mean T1 coherence time in us")

def main(argv):
    del argv  # Unused.
    logging.info(f"Initializing Drift Orchestrator with T1={FLAGS.t1_decay_mean}us")
    # ... Your Neural ODE logic here ...

if __name__ == "__main__":
    app.run(main)