requirements.txt
absl-py==2.0.0
jax==0.4.23
jaxlib==0.4.23
stim==1.12.0
pymatching==2.1.0
numpy==1.26.2
scipy==1.11.4
typeguard==4.1.5
ml_dtypes==0.3.2
opt_einsum==3.3.0